Tests/LibWeb/Text/expected/css/shadow-root-replacesync-slot-style-invalidation.txt

before replaceSync: rgb(0, 0, 0)
after slot replaceSync: rgb(255, 0, 0)
after slot replaceSync removal: rgb(0, 0, 0)
